Move Factory class util module
